requirements.txt
numpy
pandas
rdkit
scikit-learn
tensorflow-macos
transformers
tqdm
matplotlib
seaborn
jupyter
ipykernel